Fix locking
Unfortunately, fcntl locking takes the lock for the whole process --
all the threads.  So we have to use a mutex to prevent access from
other threads within the same process, while keeping the fcntl lock to
prevent interference from other processes.

(An alternate idea is to use a listening socket as a lock, but there
are probably reasons why nobody does this).

Fixes #3047